refactor: SearchDialog 컴포넌트 분리
