+++ src/ScheduleTables.tsx
@@ -5,16 +5,13 @@
 import { Button, ButtonGroup, Flex, Heading, Stack } from "@chakra-ui/react";
 import ScheduleTable from "./ScheduleTable.tsx";
 import { useScheduleContext } from "./ScheduleContext.tsx";
-import SearchDialog from "./SearchDialog.tsx";
+import { SearchDialog } from "./SearchDialog/SearchDialog.tsx";
+import { SearchInfo } from "./SearchDialog/types.ts";
 import { useState } from "react";
 
 export const ScheduleTables = () => {
   const { schedulesMap, setSchedulesMap } = useScheduleContext();
-  const [searchInfo, setSearchInfo] = useState<{
-    tableId: string;
-    day?: string;
-    time?: number;
-  } | null>(null);
+  const [searchInfo, setSearchInfo] = useState<SearchInfo | null>(null);
 
   const disabledRemoveButton = Object.keys(schedulesMap).length === 1;
